Move all tests for assets.serlo.org into one file
